requirements.txt
python-dotenv
pandas
numpy
scikit-learn
tensorflow
torch
matplotlib
seaborn
jupyter
requests
python-binance
ta
quantstats